Move config/log to system/config/log
